Add loading/error state
Summary:
While the code review is running, we need at least some kind of visual indicator in the UI. We can represent this for now using a loading spinner, and also add states for success / error cases.

This by no means should be the final state for this UI, but I think it's OK as a start.

Reviewed By: stbfb

Differential Revision: D80295715

fbshipit-source-id: 2ea7baa3f5334048f0e3a80f68d81631c062b388