Add diagnostics provider for rendering code review results
Summary:
We discussed a couple of different ways for showing the review results in VS Code. One way is to surface them as diagnostics, which will render a squiggly underneath. This works nicely because the format that we're using in the response has a severity, which can be easily represented using different colors / icons in the Diagnostics API.

That being said, I don't think this should be the primary way for showing code review comments. It's not really clear that the diagnostics are coming from Devmate code review, even when explicitly marked that way. The editor insets we use for inline comments is much nicer, and we should align with this for consistency. We may be able to use both, so we can keep this for now.

Reviewed By: stbfb

Differential Revision: D80295714

fbshipit-source-id: daf89520df8684ce5e86a628ef8978448cae30ee