Add button for initiating AI code review
Summary:
This is the start of a rough attempt to add AI code review to the ISL. The first thing we want is a way to initiate the request to Devmate from the ISL, right when the user has finished authoring code and is ready to submit it.

The original mock-ups had the review run automatically, but there are concerns about the value/capacity trade-off with this approach. For now, we'll make the trigger for code review manual.

Reviewed By: stbfb

Differential Revision: D80295718

fbshipit-source-id: f50298a6e684436840cbc3b50dbe8cbb3ead1d73