Parse code review response
Summary:
The `devmate` CLI returns a big mass of stdout, but we only care about a small percentage of it. We need to parse the review JSON object from here in order to render it properly.

I played around with this a bunch, but the most success I got was by doing the following:
1. Clean the stdout of unnecessary characters / whitespace
2. Grab the substring representing the "observation" after the `format_code_reviews_tool` runs.
3. Use a naive bracket-matching strategy to identify and parse the JSON objects.*

*I initially tried doing this using regex, but this was returning weird results if the review comments contained the substring "Review:{...}" (which happened to be the case when writing this diff). I will probably switch to a different regex approach later when we can swap in the SDK, but for prototyping purposes I think this works fine.

Reviewed By: stbfb

Differential Revision: D80295716

fbshipit-source-id: 5f3cb6bc761134c5e46740065b482418eb3fb153