Call code review using devmate CLI
Summary:
To run the code review, we need to call the correct MCP file via Devmate Platform (this is currently a substitute for ADR). The issue is that Devmate Platform isn't currently available in Typescript, so we need to call the `devmate` CLI directly.

This hooks up the necessary pieces to trigger the code review through the `devmate` CLI when the button added earlier is clicked. Currently there's no indication of when the review starts/completes besides log statements, but this will change soon.

Reviewed By: stbfb

Differential Revision: D80295719

fbshipit-source-id: 790c6f4717ce507ecff29d579753f03ada74e7ce