Add tracking for clicked smart actions
Summary: One thing we're missing that we need to have in place before going out to more users is proper tracking. We already have tracking for some smart actions, which I left untouched, but I want to also have consistent tracking for all smart actions. We should also track when users click to open the smart actions menu, regardless of whether a smart action is clicked.

Reviewed By: stbfb

Differential Revision: D80624205

fbshipit-source-id: 673f41935c9b8306e1be0b31515f61aa04521975