slapi: merge bookmarks endpoints together
Summary:
`bookmarks2` is an extended version of the bookmarks endpoint that offers extra functionalities, such as error propagation and freshness specification.

It's currently used in modern sync, but the plan is to further roll it out for all other use cases so we can have better visibility and unblock other projects like wireproto deprecation. Putting this up as a request from quark-zju for deduplicating the endpoint on SaplingRemoteApi trait.

Reviewed By: quark-zju

Differential Revision: D80527062

fbshipit-source-id: 3db4e96e90d26daa846958dc821cd3ee65899d28