hgmain: drop buildinfo feature
Summary:
The `buildinfo` feature is to emulate a buck build feature in cargo build so
the generated binary contains special symbols like `BuildInfo_kPackageVersion`
so it's possible to figure out the versions of the running binary by looking at
its symbols in a universal way.

Nowadays we use buck build on all platforms, which natively has this feature.
Let's just drop the cargo emulation.

Reviewed By: zzl0

Differential Revision: D80750191

fbshipit-source-id: c080575b82fcc89870dd8c9f608fcd99fb10fbe7