dirsync: fix mirror_data data capture
Summary:
This is subtle. But `lambda: dst_data` can cause issue if the lambda gets
accessed in the next loop iteration when `dst_data` is changed. To avoid
such issues, capture the `dst_data` in the lambda immediately so it will
not be affected by loop iteration.

Reported by: Chef-SWanger

Reviewed By: Chef-SWanger

Differential Revision: D80959221

fbshipit-source-id: ff4aa88d74db5fade834c53fd65c28b3edf49e03