slapi: retry on curl SSL_CERTPROBLEM error
Summary: This error corresponds with the message "Problem with the local SSL certificate". I saw this error in eden's logs during a failed "checkout" operation. There seems to be a temporary permission issue as the on-disk certificate is renewed, causing curl to briefly fail reading the file. Let's retry in this case since it is very hard to be certain a vague certificate error is permanent when it is possible certificates are being renewed on disk.

Differential Revision: D80973971

fbshipit-source-id: 83af4fc8a20ee174659789582478f1e9038b4f7a