rename some benchmarks
Summary:
This better reflects the new reality of the benchmarks. The existing getxattr benchmark was intended to measure syscall overhead. The get_sha1 benchmark was intended to measure getxattr/getSHA1() perf, but now it supports measuring the performance of several different attributes (via thrift and via FS).

Therefore, we rename everything to better reflect their uses.

Reviewed By: genevievehelsel

Differential Revision: D80856111

fbshipit-source-id: 094e364650d9a041a1e9390a354cb11b0deaa8b7