configloader: add "local_load()" entry point that skips remote fetch
Summary:
Add a new entry `hg::local_load()` entry point that only uses the locally cached remote config, erroring out if the config is not present.

This can be used for relatively lightweight decision making that wants to make use of all the config features, but doesn't want to wait for a remote fetch.

It is implemented by using the FbConfigMode::OfflineJsonOverride config mode, pointing to the local cache file location. If the file doesn't exist, it will error out (previously it would panic due to an `unwrap()`). I changed OfflineJsonOverride to contain a PathBuf instead of a String since going from PathBuf to String is annoying.

Reviewed By: quark-zju

Differential Revision: D80994000

fbshipit-source-id: 3322ca5afcee88d5076b1b1388b763fc4509afdc