chg: allow configuring cgroup namespace behavior
Summary:
In a previous diff we started namespacing the chg socket path by cgroup id to avoid chg accidentally executing sapling command in unexpected cgroups. This, however, has the potential for spawning a lot of chg servers if there are a lot of cgroups running sl commands.

Support a new chg.cgroup-regex config which is regex to match cgroup names we want to namespace. First, this allows disabling the namespacing altogether by giving ".*" as a regex. Second, it allows us to specify which cgroups should get namespacing applied to them. For example, if you write "user.slice" to the file, the socket will be namespaced by cgroup id only if the current processes cgroup name includes "user.slice".

If chg.cgroup-regex is set but doesn't match the current cgroup, chg is not used. This is intended to be used with "random" short lived cgroups we don't want to spawn a long lived chg for.

This commit makes use of the new `configloaded::hg::local_load()` function to load the sapling config, but avoid remote config fetches (to be faster). If there is no cached config, we will simply not use chg for now (the cached config will be written by the next sl command).

The config loading takes around 2ms in my testing. This is not nothing, but being able to configure chg using the sapling config (particularly the dynamic config) is a huge quality-of-development improvement, I'm going to say it is okay.

Reviewed By: quark-zju

Differential Revision: D80947096

fbshipit-source-id: a71b9f5fe99bc5be156436d78a6eea1bbcdcc922