chg: clean up socket on SIGTERM
Summary:
Set up SIGTERM (and SIGINT) handler to raise an exception, which will trigger the `finally` block that cleans up the socket file.

I think we lost this cleanup when we switched to the Rust signal handling. The Rust handlers are not set up for chgserver, so SIGTERM was exiting the chg server immediately.

It wasn't a problem in practice because the sockets live under `/var/run/user/<uid>` which gets cleaned up on reboot (?).

Reviewed By: quark-zju

Differential Revision: D80944400

fbshipit-source-id: d766200283d6ab679a800534afeb7d477e528757