indexedlog: enable btrfs compression when btrfs_compression=true
Summary:
When the log is in btrfs_compression mode, set the "btrfs.compression" property on the log file when opening it. This is to try to ensure btrfs compression is actually enabled.

Note that the file must be in write mode and must not already be mmap'd (or you can't set the attribute). In general there seem like a lot of cases where you can't set the attribute, so we ignore errors.

Reviewed By: quark-zju

Differential Revision: D80317903

fbshipit-source-id: 0ca59ab2d58b0dd3777f2e605b692277cfe84270