indexedlog: add btrfs compression awareness to RotateLog
Summary:
Add some logic so that log rotation is based on the "phsyical" log size on btrfs (i.e. compressed size) when btrfs_compression mode is enabled.

The main implication is that logs will be able to hold more data per-log. Previously if we set the max-log-size to 1GB, we would rotate when the apparent size was 1GB, but the physical size might have only been 200MB. If we end up enabling this mode, we will probably want to tune hgcache sizes to be a bit smaller.

Reviewed By: quark-zju

Differential Revision: D80317892

fbshipit-source-id: e8886b7851e5c958fb8e8c6909d5f1a84bd9d4d5