walkdetector: only track recursive counts for large directories
Summary:
I noticed we can spend a good chunk of CPU in TreeNode::set_metadata updating the vec containing "rollup" counts for descendant directories. The main purpose is to track really big directories, so tweak things so only "important" metadata is tracked in the rollup counts. This saves futzing with the vector unless actually important. With current defaults, previously it was tracking the recursive metadata for descendant directories with 3 or more entries, and now it will only track directories with 100 children.

I also moved TreeNode's total_dirs and total_files values back to direct struct fields instead of the first element in the vector to avoid allocating vectors for the common case of not-really-big-directories.

Note that some edge cases are not handled, such as a large directory becoming a small directory. Nothing bad will happen per-se, but the walk detector will continue to think the directory is big when making walk decisions.

Reviewed By: zzl0

Differential Revision: D81265737

fbshipit-source-id: 4385a3f4733c9cfdd59bfc8bc23a9bfcc930d1f0