http-client: optimize cbor decoding for small requests
Summary:
Optimize CborStream to avoid copying the first body chunk. This eliminates another copy of the data when fetching a single file/tree that arrives in a single chunk from curl (the chunk size seems up to 16k).

The CborStream was generic across a stream of AsRef<[u8]>, but in practice it was only Vec<u8>, so let's just drop that generic param.

Reviewed By: quark-zju

Differential Revision: D80317914

fbshipit-source-id: ddbf47b04c731a76f310f49afa3206868506e92c