revisionstore: consolidate LFS stuff into the LfsClient
Summary:
The FileStore confusingly contained an lfs_remote (which was actually LfsClient, not LfsRemote), lfs_local, and lfs_cache. However, the LfsClient also contained the local and cache stores. Consolidate things by making the FileStore only hold the LfsClient (and reach inside when it wants the cache/store). This isn't that great either, but less confusing.

The main thing I actually want is a unified `flush()` method that is called on the LfsClient object (I want to add something new that happens on flush()).

Reviewed By: quark-zju

Differential Revision: D80763639

fbshipit-source-id: cc11c19d44bef9e2c227df560bab7a993fc7a902