revisionstore: fix LFS download parallelism
Summary:
Revert D56073386, getting rid of lfs.max-batch-size.  I don't think I fully understood this code when I made that change.

The LFS download code first does a batch fetch to get the oids of all the files, and then creates an async stream to download the files. The stream is run through `buffer_unordered()`, so the number of concurrent downloads is already limited.

In D56073386 I chunked up the batch request to get the oids which might be helpful in some rare cases you fetch 1,000,000 LFS files in a single request, but it had an unintended side effect of making each chunk of LFS downloads have to wait for the entire previous chunk to finish, thus slowing things down a lot (when fetching >100 LFS files in a single request).

I noticed this during "sl prefech" where there was an obvious pattern of fast->slow->fast->slow.

Reviewed By: quark-zju

Differential Revision: D80666105

fbshipit-source-id: 7d029c3f205a6c14b593765dd1b07d2403fa9eea