revisionstore: add LFS streaming inserter
Summary:
Add a streaming chunk inserter that allows inserting LFS chunks into the LFS blob store as you go instead of gathering the entire LFS blob in memory. I will use this in upcoming diffs to reduce memory usage when fetching LFS files.

The one tricky thing is that we currently verify LFS content hash before inserting. To achieve that behavior with the streaming inserter, we delay inserting the final chunk until we have verified the content hash. This means some corrupt chunks might be inserted, but when reading blobs, we treat missing chunks as an overall `None` since it is always possible to be missing chunks. Further, this approach assumes that subsequent attempts to download and write the LFS file won't try to re-use chunks already in the store, which is currently the case.

Note: we choose to verify SHA256 at write time instead of read time for performance reasons (read is more frequent and SHA256 is slower than hashes like xxhash).

Another optimization in this streaming approach is we can re-use the serialization buffer to avoid another copy of each chunk.

Reviewed By: quark-zju

Differential Revision: D80317883

fbshipit-source-id: 2b8ed4565fc62e8821bc77e746bdbe713a14dd7d