scmstore: don't buffer LFS files in memory
Summary:
Use the newly improved LfsClient::fetch_batch which streams LFS content into the indexedlog cache instead of buffering in memory. After fetching files, we flush the indexedlog cache (to get any LFS chunks out of the indexedlog's in-memory buffer), and then read from the indexedlog cache (which after recent improvements returns a folly IOBuf chaining the mmap backed chunks together).

We also make use of the RotateLog's new "consistent reads" mode so we can be confident LFS entries aren't going to rotate out of the cache between writing and reading.

I added an experimental.lfs-buffer-in-memory config option which can be enabled to revert to the previous behavior.

Reviewed By: quark-zju

Differential Revision: D80317908

fbshipit-source-id: 89b3666a40ea46592642a54077b5083b03bda4b3