revisionstore: add a "limited" pool for sharing buffers
Summary:
Add a LimitedBufferPool type which is a pool to share buffers, but is also a "limiter" that will only hand out up to N buffers concurrently. The buffers we make when downloading LFS files are so massive (~16MB), they should be re-used lest our heap explodes with concurrent downloads.

The pool allows callers to put used buffers back into the pool, but has a safety mechanism to free up a slot even if the caller forgets to put the buffer back.

Reviewed By: quark-zju

Differential Revision: D80666099

fbshipit-source-id: 6484feff1ec16baa536265ce74694df1e4e32601