http-client: avoid unnecessary large allocation decoding cbor
Summary: Avoid using `Bytes::into_vec()` since that will copy the entire Bytes if it can't extract the Vec. The Bytes could be very large if we saw a large file. Instead, we allocate a new buffer with capacity for the current threshold (which should be strictly smaller than the size of the Bytes).

Reviewed By: quark-zju

Differential Revision: D80317877

fbshipit-source-id: 5b78ac7e2f08f3b06a7adb6abf94076bd7c37a2b