revisionstore: make LfsClient::fetch_batch stream to cache
Summary:
Previously LfsClient::fetch_batch just called LfsRemote::fetch_batch, ignoring the client's cache. Now it uses the new streaming download mechanism to write the downloads directly to the LfsClient's shared blob store (i.e. cache). This way the blobs' content doesn't need to be held in memory.

The callback was changed to a simpler "this-hash-is-done" callback since the full content is not available due to the streaming nature.

Nothing currently calls LfsClient::fetch_batch, so it is safe to change it.

Reviewed By: quark-zju

Differential Revision: D80317887

fbshipit-source-id: eb4150c76ddedc0f970cb76290900755fbce59e5