indexedlog: support read-after-write consistency in RotateLog
Summary:
Add a new `log.with_consistent_reads()` method that guarantees read-after-write consistency for this RotateLog instance while the returned guard object is alive. This is useful if you want to quickly flush data to the RotateLog to avoid memory usage, but are afraid of the possibility that the data gets rotated out before you read it back it. For example:

```
for id in ids {
  rotate_log.append(id, download_big_data(id));
}

// Flush data into log so we don't hold big data in memory.
rotate_log.sync();

// Read data back out as slices from mmap'd indexedlog.
// XXX data is not guaranteed to be present anymore!
return ids.iter().map(|id| rotate_log.get(id));
```

In particular, I want to use this consistent read mode when fetching LFS files (following the above pattern).

I decided to implement the consistency by "pinning" Log objects that would be rotated out in a separate Vec stored on the RotateLog. All the rotation (and deletion) still happens like normal on disk, we just keep the log and index data alive through the Log objects. And, we don't have to buffer any extra data in memory (but we will temporarily use extra disk space until the pinned Logs are dropped).

Reviewed By: quark-zju

Differential Revision: D80317891

fbshipit-source-id: de7fba27a3c394367770ae4b9bbf149ebbc7ba2d