indexedlog: add optimized Log::append_direct method
Summary:
Add a new `append_direct` method that takes a callback, allowing the caller to write log data directly the Log's internal in-memory buffer. This helps the caller avoid copying the data.

For example, Log insertions often look like:

```
let entry = Entry {
  metadata,
  data,
};

// Allocates and copies `data`.
let serialized: Vec<u8> = entry.serialize();

// Internally Log copies `serialized` into the memory buffer.
log.append(&serialized);
```

Now they can look like:
```
let entry = Entry {
  metadata,
  data,
};

// Serialize directly to the Log's buffer, skipping intermediate allocation and copy.
log.append_direct(|buf| {
  entry.serialize_to(buf)
});
```

In order to not change the Log's serialized format, I made `append_direct` first call the callback with an object that only records the number of bytes written. This allows the `append_direct` to know the exact data length, which is recorded as a varint (so you can't "reserve" space since you don't know how many bytes will be needed).

In a previous iteration I added a new flag to indicate the data length was recorded in a fixed length field, but this is a bit annoying since it is a breaking format change. The two-phase approach to detect the exact length doesn't seem to add much overhead, so I went with that. It also saves a significant amount of space vs the fixed 8 byte length for small entries (i.e. up to 7 bytes per entry, which is a lot of each entry is small).

Reviewed By: quark-zju

Differential Revision: D80317899

fbshipit-source-id: e686a435d33dadc69cf8d3c930b191dde7b8696b