mysql: Expose Freshness argument in Bookmarks facet get method
Summary:
## What
Update the `get` method in the `Bookmarks` facet to take a `Freshness` argument (same as `list` method takes).

## Why?
Please read the summary of D81212709. It has details about the problem we're solving.

The TL;DR is that when fetching a bookmark value through the Bookmarks facet we're **always** reading from the primary. This is creating a huge number of connections and leading to us hitting the 4K limit of connections to the DB.

## This diff
During the Thursday discussion, RajivTS brought up a good point: changing the default behaviour for all the `get` calls will likely break a few things that **actually need the most recent bookmark value and should read from master**.
This was later proven by the broken tests in V1 of D81212709 (which was taking that approach).

The 2nd approach is: expose a `Freshness` argument in the get method **so each callsite can specify how fresh they need the bookmark to be**.

Please note that I added the exact two lines to all the callsites:

```
// TODO(T236130401): confirm if this needs read from primary
bookmarks::Freshness::MostRecent,
```

this is to **facilitate searching and updating callsites in bulk**.

This diff should be a no-op, because all callsites will pass `MostRecent`.

In the next diff, D81212709, I'll update the `SqlBookmarks` logic to use the freshness argument **based on a JK value**.

After that, I'll put up diffs reviewing and updating each callsite.

I will very likely need some help from a few people to decide what should be done in a few callsites.

## RFC
I was thinking if in this diff I should make the argument **optional** instead of required.

Advantage of making it required: it will be explicit in **all the callsites** what's happening, instead of needing to track the code to find the default behaviour when `None` is passed.
However, when new callsites are added, users adding these calls are likely to pass `MostRecent` without fully understanding the tradeoffs. I added doc comments to avoid that though...

If we make it optional, we can more easily change the default behaviour in the future by changing only the `get` function in `SqlBookmarks`.

Differential Revision: D81238781

fbshipit-source-id: 8fa78bc12f41e88a406795df96bb869dd6a78b9d