Stop reading bookmarks from XDB primary
Summary:
## What
Update `SqlBookmarks` logic to read bookmark values from a READ replica, instead of the primary.

## Why
**TL;DR:** this is creating **a lot** of unnecessary connections to the primary of `xdb.mononoke_production` and making us hit the 4K limit of the connections to it.

For a long time we've been reading bookmark values directly from the primary. Very likely to ensure that we get the up-to-date value.

Before I explain the most likely contributor to this problem, let me give some context on the flow of bookmark reads in Mononoke, SCS and Mononoke Git.

- When a bookmark value needs to be resolved, this will be done by calling [Mononoke API's `resolve_bookmark`](https://fburl.com/code/hcbb0257).
- This will try to read the bookmark value from the "Warm Bookmark Cache", which, depending on the service, might be the original WBC (in SCS), the Bookmark Service (Mononoke) or pub-sub version of WBC (Mononoke Git).
- **If the bookmark is not found there, it will then try to get its value directly from the `bookmarks` facet**.
- The bookmarks facet is a `CachedBookmarks` ([code](https://fburl.com/code/0402r0kt))
- This `CachedBookmarks` is a wrapper that caches bookmarks **for list operations**, not for get (i.e. read specific bookmark's value). See [the code](https://www.internalfb.com/code/fbsource/[296d8d5dc2e302a6b8b89954bfc2b9d291ee100b]/fbcode/eden/mononoke/bookmarks/src/cache.rs?lines=347).
- It's inner bookmarks is the `SqlBookmarks`, which is what actually makes queries to the DB. This is what [queries the primary directly for `get` operation](https://fburl.com/code/qzvowvo0).

I believe that the biggest problem is coming from **trying to read bookmarks that don't exist. All of these attempts will lead to a query directly to the primary.**

So let's change this behaviour of reading from the primary to see how much the number of connections will decrease.

## Notes
**Bookmark creation**
One edge case that we need to handle with this change is bookmark creation. Users might create a bookmark and try to read its value right after that, so this change might lead to errors saying that the bookmark doesn't exist.

To avoid this I'm working with the MySQL team to see if they offer something on their transactions API like "commit_and_replicate", which would ensure that the change has been replicated to all the replicas before returning control to us.

In the meantime, we can sleep for 1s to ensure that because our staleness is consistently below this value: https://fburl.com/canvas/89d1dfm6

**Most important callsites to review and update**
After shipping this diff, all the callsites that call `get` from the `Bookmarks` facet directly will still be querying the primary replica.
**However**, all the calls that go through `resolve_bookmark` will use the provided Freshess arg and the SCS method `repo_resolve_bookmark` will pass `Freshness::MaybeStale`.

I checked that [it's responsible for ~25% of the reads from master](https://fburl.com/canvas/xrnh0av2), so just shipping this diff and enabling the JK should be a big improvement!

It looks like there are [other methods](https://fburl.com/scuba/mononoke_scs_server/be0uqjsl) that might also be causing problems, so I'll look into them as well. These are probably the ones triggering reads to the mapping tables, which are the biggest contributors after the `SelectBookmark` query ([query](https://fburl.com/scuba/mononoke_sql_telemetry/q6cdd6dg))

Reviewed By: andreacampi

Differential Revision: D81212709

fbshipit-source-id: 85ff3b99fcd60eca0965986602ef96dfe3f74b88