hash -> id where applicable
Summary:
# This diff

Attempts to replace all instances of "hash" to "id" when we use "hash" to refer to ObjectIDs/RootIDs. See context below for more information on the principles I followed for the rename

# Context

Many moons ago we switched the identifiers that we use to represent source control objects (see D31316477 for some context). The gist is that we switched from calling identifiers "hashes" (or "proxy hashes") and moved to calling object identifiers "Object Ids".

This was a smart move, since it separated the internal representation that EdenFS uses to identify objects (Object Ids) from the external representation that a particular BackingStore might use (i.e. Sapling hashes, git hashes, etc). In other words, it made object identifiers opaque and prevented (or at least tried to prevent) BackingStore implementation details from leaking into core parts of EdenFS.

The downside of the switch is that the code is now littered with "hash" variable names, field names, and types (most of which refer to Object Ids, but not all!).

Therefore, variable/field/type names aren't useful for figuring out what some code refers to; one must look at the type/variable definition to understand if an ObjectID is being used, or if a Sapling/Git hash is being used (or blake3, or sha1, etc).

This diff stack aims to end that confusion as much as possible by moving to less ambiguous naming for types. General principles I followed in this stack:

1) When "hash" refers to ObjectID, switch to "id"
2) When "hash" refers to RootID, switch to "id" or "scmId", or "root"
3) When "hash" refers to ProxyHash, keep it as is.
4) When "hash" refers to SHA1, Blake3, or other hash type, keep it as is.
5) When "hash" is used for a Thrift type/field, keep it as is, since renaming a Thrift type/field is unsafe (we should maybe follow-up here).

Reviewed By: muirdm

Differential Revision: D81203399

fbshipit-source-id: 8d65d2e463f1bf13a288e58a5d87a6a72c7f30b7