FileInode: fix materialized inode tests
Summary:
For 7 years, these tests didn't actually test what they said claimed to be testing. The materialization logic never ran because the write() futures were created but never executed. Therefore, they were basically duplicates of the non-materialized versions of the test.

This diff fixes the problem by forcing the future to be executed via .get(). This is verified in the next diff, as we assert that the file is materialized with a newly added isMaterialized() method.

Reviewed By: muirdm

Differential Revision: D81203391

fbshipit-source-id: 1d54ddd37f714c8d32c6bce6617bf14735045e4c