VirtualInode: use collectAllValid in getAttributesFromFilesV2
Summary:
# This diff

Uses the newly added collectAllSafe() method to lazily evaluate the attribute futures for only the requested attributes. Previously we used dummy/placeholder values for attributes that were not requested, causing the code to be cluttered/confusing. Now, we can neatly initialize all potential futures as "empty", and then conditionally collect the futures that are non-empty with collectAllValid()

# Context

Addresses some feedback on D81190336

Reviewed By: genevievehelsel

Differential Revision: D81343703

fbshipit-source-id: 04279cc3ae20f642cb1a3da792451c48fa413f2a