Include concurrency parameter in commit graph preloading
Summary:
The `GitCommitGraphPreloader` chronos job takes about 8 hours to complete in every iteration even if there are not that many changes between each run. This is because it processes only 10 repos at once and with 5000+ repos to handle, this can end up taking a long time.

This diff adds a concurrency parameter that can be used to speed this up and defaults its value to 100. If the chronos job is still too slow, this can be cranked up even further

Reviewed By: YousefSalama

Differential Revision: D81439832

fbshipit-source-id: d5abd0cf61d7fc9218e3ed7b5c7e9835705330b3