mysql: Remove sleep after bookmark creation
Summary:
Remove part of the changes from D81212709, which broke a Mononoke Git test and are not actually needed.

These were introduced because of the original implementation of D81212709, which would change the default behaviour of reading from master to **all calls of Bookmarks::get**.

The newer implementation, which was actually shipped, allows each callsite to specify the freshness of the Bookmark, so calls that really need it will keep reading from master.

The problem right now is that when the JK passes, we introduce an artificial delay of 500ms during bookmark creation, which is what is breaking the test.

Reviewed By: RajivTS, lmvasquezg

Differential Revision: D81436471

fbshipit-source-id: a0b60e181e4b456acfa8208773ea3f80ddac5c51