git_delta_manifest_v3: refactor GitDeltaManifestOps trait to make the path part of the entry
Summary: In GDMv2, the entry path is formed by the edges of the sharded map (and also in the deltas which is a waste of space). In GDMv3, the path is stored inside of the entries instead. This diff refactors the `GitDeltaManifestOps` traits to make the path part of the abstract entry representation. To make this work the implementation of `GitDeltaManifestEntryOps` for `GDMV2Entry` is now implemented for the tuple of `(MPath, GDMV2Entry)` instead.

Reviewed By: RajivTS

Differential Revision: D81142211

fbshipit-source-id: 42f22db9b54d70535b230b614e5a3c409736ddaa