mysql: Keep reading bookmarks from primary in git code
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff
This seems to be derivation code, so like D81446419, let's leave it as is?

cc RajivTS pierrec YousefSalama if you think differently, i.e. if you think there's an opportunity to improve performance by reading from a replica.

Keep in mind that if you can read from a replica, this can **significantly reduce latency**, as I am seeing in [an ongoing experiment](https://fb.workplace.com/groups/3112141815538647/permalink/24488073400852179/), where the impact on latency is big for some regions: https://fburl.com/canvas/8mwajzly

Differential Revision: D81447238

fbshipit-source-id: 4684987b8bc3d6a88e81bdeb2439a2bcc3bb5f86