mysql: Keep reading bookmarks from primary in megarepo code
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff
Like cross-repo sync in D81446491, let's leave this untouched because it involves syncing across repos.

Reviewed By: andreacampi

Differential Revision: D81447234

fbshipit-source-id: 7e3ea96adf0d0824c06664f23c7443961e9c1b5e