manifest: handle replacement manifests for added trees
Summary:
If the parent of the subtree copy destination didn't exist, then we fail to perform the comparison as we think it's missing.

This is caused by not handling the replacment in the "Added" case.  If a subtree copy destination is in a directory that didn't appear in the parent, then the parent directory of the destination appears as "added" in the diff.  We still need to process manifest replacements here, as this is how we will compare the child directory against its subtree copy source.

Reviewed By: zzl0

Differential Revision: D81417262

fbshipit-source-id: de96b0ddf2f1718556a340323a79415fd8a91890