Adds the rust client equivalents for StatechangeNotification.
Summary:
# This Diff
Adds client equivalents for StatechangeNotification.

A state change consists of either StateEntered or StateLeft, which contain the name of the state being asserted/deasserted and the sequence number for differentiating between multiple enter/exit events.

## Why move notifications into the mount

The primary justification for doing this is to add guranteed temporality into
Previously, it was possible to miss entire state-enter/leave events if they completed while eden was in the process of checking the lock.

By inserting the lock events into the change stream, you can get clearly delineated borders for when states were entered or exited.

# Technical Details

# Discussion Points

Reviewed By: muirdm

Differential Revision: D75562824

fbshipit-source-id: 8e31fbcfd639e5be11eae26926978b2d8ed9005f