Read includeStateChanges value in changesSinceV2
Summary:
## This Diff
Read the new thrift field includeStateChanges.
Log the value.
If present, add values from the state folder.

Currently, this has no effect, because nobody is setting the includeStateChanges flag yet.
If it were to be set, there would be additional changes reported when state changes would appear in a relative root.

Reviewed By: muirdm

Differential Revision: D79133616

fbshipit-source-id: 7ac5bc5c98bb1f05de4ff77dd1913888608c20fd