Implements returning state changes from ChangesSinceV2
Summary:
# This Diff
Implements checking the events created inside the notifications directory
The state event is triggered by the creation or deletion of a file named <state>.notify in an appropriate subfolder.

This logic is intended to augment the checking done in edenfs-notifications-state.
The checking for a crashed state will still be done in that file.

Reviewed By: muirdm

Differential Revision: D75562838

fbshipit-source-id: ad4adf3dcd2e1df4c117de1d0e5b41f7d34fc239