Overlay: rename readExistingOverlay
Summary:
# Context

I'm preparing for a project that aims to significantly improve EdenFS write/checkout performance by removing quadratic complexity from the Overlay layer. During preparation, I'm reading through the Overlay code to understand what the limitations are and where we can make improvements. I'll try to fix issues and make small improvements where possible during my preparation.

# This diff

The naming of this method is confusing -- it makes it seem like the overlay is read/parsed. In reality, it's just performing a sanity check on the overlay to verify that the current EdenFS daemon knows how to interpret existing overlay data. Rename the method to reflect it's actual purpose.

Reviewed By: kavehahmadi60

Differential Revision: D80631337

fbshipit-source-id: 791d0a0d3477738d48417578da22537b4c565532