filenodes: remove scm/mononoke:filenodes_disable_master_fallback JK
Summary:
The `scm/mononoke:filenodes_disable_master_fallback` JK allows us to disable fallback to master for filenodes lookup.  However, it's flawed in a couple of ways:

* If enabled, and if the filenode is not found on a replica, we return `Disabled` rather than `None`.  Despite the intention of the design of `FilenodeResult::Disabled` there are too many places in Mononoke which don't actually handle it (and call `do_not_handle_disabled_filenodes()` instead).  This means *any* filenode miss, even a legitimate one, becomes a hard error.
* If enabled, then things that *are* sensitive to replication lag suddenly start failing.  Landing is one such thing, as it checks that filenodes are derived for recent commits in `master`, and these are very likely to not be replicated to a db replica yet.

As a result, this JK is not safe to ever enable, so we should just remove it.

As an alternative, we should improve the filenodes API to allow the client to specify staleness, and have replica-only fetches be opt-in.

Reviewed By: chiara125

Differential Revision: D81514033

fbshipit-source-id: 850a09ccfd344f4dbd6a4a1b728c9764be83b1cb